addressing comments from review, added conftest module for fixtures
